usb: dwc3-msm: Don't issue power supply change for invalid charger
This is constantly causing wakeups and general havoc/instability, as it
constantly issues a power supply change for the same power supply type
(DWC3_INVALID_CHARGER).

Constantly issuing a power supply change for an invalid charger doesn't
make sense, so don't do it in order to fix the USB madness.

Signed-off-by: nancyjudge <taitowatashiwa3@gmail.com>